Created separate NavigationStack component
Want to keep backward compatibility so need to keep NavigationMotion around. Too complicated, for no gain, to change NavigationMotion to work in 2 ways - motion or keyframes. So created separate NavigationStack component - much better name that NavigationMotion anyway, especially with making Web and Native match. Now the crumbStyle and unmountStyle props match too makes sense to use the same component name.
Also means can release (and even update docs?) without doing shared element first
